refactor: unify assets and main styles/scripts
