+++ archives/2024/index.html
@@ -65,7 +65,7 @@
                       
                         
                               
-<link rel="stylesheet" href="/css/style.css">
+<link rel="stylesheet" href="/css/main.css">
