--- assets/css/style.css
@@ -1,173 +1,0 @@
-/* style.css - 全站樣式 | Global Styles */
-/* ================================================ */
-/* Variables 變數 */
-:root {
-  --color-bg: #050505; /* 碳黑背景 | carbon black */
-  --color-primary: #00FF9F; /* 霓虹綠主色 | neon green */
-  --color-secondary: #19F6E8; /* 酸性青次色 | acid cyan */
-  --color-accent: #C9AE5D; /* 電路金點綴 | circuitry gold */
-  --font-mono: 'IBM Plex Mono', monospace;
-  --font-body: 'Inter', sans-serif;
-}
-
-/* Reset 初始化 */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: var(--font-body);
-  background: var(--color-bg);
-  color: var(--color-secondary);
-  line-height: 1.6;
-}
-
-/* 通用排版 | Layout */
-.container {
-  width: 90%;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-/* 導覽列 | Navigation */
-.nav {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  background: rgba(5, 5, 5, 0.8);
-  color: var(--color-primary);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 2rem;
-  z-index: 1000;
-}
-
-.nav__logo {
-  font-family: var(--font-mono);
-  font-size: 1.2rem;
-}
-
-.nav__list {
-  list-style: none;
-  display: flex;
-  gap: 1rem;
-}
-
-.nav__link {
-  color: var(--color-secondary);
-  text-decoration: none;
-  position: relative;
-}
-
-.nav__link:hover {
-  color: var(--color-primary);
-}
-
-/* 頁尾 | Footer */
-.footer {
-  text-align: center;
-  padding: 2rem 0;
-  color: var(--color-secondary);
-}
-
-/* Hero 區域 */
-.hero {
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  overflow: hidden;
-}
-
-.hero__canvas {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: -1;
-}
-
-.hero__title {
-  font-family: var(--font-mono);
-  font-size: 2rem;
-  color: var(--color-primary);
-}
-
-/* 內容區 | Main content */
-.section {
-  padding: 6rem 0 2rem;
-}
-
-/* 卡片 | Cards */
-.card {
-  background: #0b0b0b;
-  border: 1px solid var(--color-secondary);
-  padding: 1rem;
-  margin: 1rem;
-  border-radius: 4px;
-  transition: transform 0.3s ease;
-}
-
-.card--flip {
-  transform: rotateX(90deg);
-  opacity: 0;
-}
-
-.card--show {
-  transform: rotateX(0);
-  opacity: 1;
-}
-
-/* 工具格 | Toolkit grid */
-.grid {
-  display: grid;
-  gap: 1rem;
-}
-
-@media (min-width: 600px) {
-  .grid--toolkit {
-    grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
-  }
-}
-
-/* 光圈游標 | Cursor ripple */
-.cursor {
-  position: fixed;
-  top: 0;
-  left: 0;
-  pointer-events: none;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background: rgba(0, 255, 159, 0.2);
-  transform: translate(-50%, -50%);
-  transition: transform 0.1s ease;
-  z-index: 1000;
-}
-
-/* 連結 glitch */
-@keyframes glitch {
-  0% { transform: translate(0); }
-  20% { transform: translate(-2px, 2px); }
-  40% { transform: translate(2px, -2px); }
-  60% { transform: translate(-2px, 2px); }
-  80% { transform: translate(2px, -2px); }
-  100% { transform: translate(0); }
-}
-
-.nav__link:hover {
-  animation: glitch 0.3s;
-}
-
-/* Typography example for headings */
-h1, h2, h3 {
-  font-family: var(--font-mono);
-  color: var(--color-primary);
-}
-
