+++ toolkit.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Toolkit - White Hat Hacker</title>
   <link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Mono:wght@400;700&family=Inter:wght@400;700&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="assets/css/style.css">
+  <link rel="stylesheet" href="css/main.css">
 </head>
 <body>
   <nav class="nav">
@@ -32,6 +32,6 @@
   <footer class="footer">
     <p>&copy; 2024 White Hat Hacker</p>
   </footer>
-  <script type="module" src="assets/js/main.js"></script>
+  <script src="js/main.js"></script>
 </body>
 </html>
